move options stuff from shared to a new options package
